chore(test): move shared test helpers to root test package
Address review feedback on #701.

Move `kubeflow/common/testing.py` to `test/testing.py` since it is an
internal test helper rather than part of the shipped package, and update
all test modules to import from the new location.

Remove the `kubeflow/trainer/test` and `kubeflow/spark/test` compatibility
shims; these are internal test modules with no external contract. The
`SPARK_CONNECT_*` constants move to
`kubeflow/spark/backends/kubernetes/backend_test.py`, their only consumer.

Also drop the redundant empty `kubeflow/spark/tests` directory, add `test`
to ruff's known-first-party list so imports sort consistently, and point
`AGENTS.md` at the new helper location.
