+++ AGENTS.md
@@ -37,7 +37,7 @@ docs/                            # Kubeflow SDK documentation
 examples/                        # Kubeflow SDK examples
 hack/                            # Scripts to manage CI/CD and installation
 proposals/                       # Kubeflow Enhancement Proposals (KEPs)
-test/                            # Top-level end-to-end tests
+test/                            # Top-level end-to-end tests and shared test helpers (testing.py)
 kubeflow/                        # Main Python package
 ├── common/                        # Shared utilities, types, and constants across all projects
 │
@@ -50,8 +50,7 @@ kubeflow/                        # Main Python package
 │   │   └── localprocess/              # Subprocess backend for quick prototyping
 │   ├── constants/                   # Common trainer constants and defaults
 │   ├── options/                     # Backend configuration options (KubernetesOptions, etc.)
-│   ├── types/                       # Common trainer types (e.g. TrainJob, CustomTrainer, BuiltinTrainer)
-│   └── test/                        # Shared test fixtures (common.py)
+│   └── types/                       # Common trainer types (e.g. TrainJob, CustomTrainer, BuiltinTrainer)
 │
 ├── optimizer/                     # Kubeflow Optimizer
 │   ├── api/                         # OptimizerClient - main user interface
@@ -178,7 +177,7 @@ uv run pre-commit run --all-files           # Run all hooks
 - Every new feature or bugfix MUST be covered by unit tests
 - Unit tests: `kubeflow/trainer/**/*_test.py` (no network calls allowed)
 - Use `pytest` with `TestCase` dataclass for parametrized tests (see `kubeflow/trainer/backends/kubernetes/backend_test.py` for the reference pattern)
-- See `kubeflow/trainer/test/common.py` for fixtures and patterns
+- See `test/testing.py` for shared fixtures and patterns
 
 ### 4. Security
