+++ kubeflow/spark/api/spark_client_test.py
@@ -18,7 +18,6 @@ from unittest.mock import patch
 
 import pytest
 
-from kubeflow.common.testing import FAILED, SUCCESS, TestCase
 from kubeflow.common.types import KubernetesBackendConfig
 from kubeflow.spark.api.spark_client import SparkClient
 from kubeflow.spark.options import Labels
@@ -27,6 +26,7 @@ from kubeflow.spark.types.types import (
     FuncJob,
     SparkJob,
 )
+from test.testing import FAILED, SUCCESS, TestCase
 
 
 @pytest.mark.parametrize(
