--- kubeflow/spark/test/common.py
@@ -1,32 +1,0 @@
-# Copyright 2025 The Kubeflow Authors.
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-"""Shared test utilities and types for Kubeflow Spark tests.
-
-Common helpers have moved to kubeflow.common.testing.
-"""
-
-from kubeflow.common.testing import (  # noqa: F401
-    DEFAULT_NAMESPACE,
-    FAILED,
-    RUNTIME,
-    SUCCESS,
-    TIMEOUT,
-    TestCase,
-)
-
-# SparkConnect states for mocking
-SPARK_CONNECT_READY = "spark-connect-ready"
-SPARK_CONNECT_PROVISIONING = "spark-connect-provisioning"
-SPARK_CONNECT_FAILED = "spark-connect-failed"
