+++ kubeflow/trainer/backends/kubernetes/backend_test.py
@@ -32,14 +32,6 @@ from kubeflow_trainer_api import models
 from kubernetes import client
 import pytest
 
-from kubeflow.common.testing import (
-    DEFAULT_NAMESPACE,
-    FAILED,
-    RUNTIME,
-    SUCCESS,
-    TIMEOUT,
-    TestCase,
-)
 from kubeflow.common.types import KubernetesBackendConfig
 from kubeflow.trainer.backends.kubernetes.backend import KubernetesBackend
 import kubeflow.trainer.backends.kubernetes.utils as utils
@@ -59,6 +51,14 @@ from kubeflow.trainer.options import (
     TrainingRuntimeSpecPatch,
 )
 from kubeflow.trainer.types import types
+from test.testing import (
+    DEFAULT_NAMESPACE,
+    FAILED,
+    RUNTIME,
+    SUCCESS,
+    TIMEOUT,
+    TestCase,
+)
 
 NOT_FOUND = "not_found"
 FORBIDDEN = "forbidden"
