+++ kubeflow/trainer/backends/kubernetes/utils_test.py
@@ -20,10 +20,10 @@ from unittest.mock import MagicMock, patch
 from kubeflow_trainer_api import models
 import pytest
 
-from kubeflow.common.testing import FAILED, SUCCESS, TestCase
 import kubeflow.trainer.backends.kubernetes.utils as utils
 from kubeflow.trainer.constants import constants
 from kubeflow.trainer.types import types
+from test.testing import FAILED, SUCCESS, TestCase
 
 # --------------------------
 # Test Helpers
