+++ kubeflow/trainer/backends/localprocess/backend_test.py
@@ -20,7 +20,6 @@ from unittest.mock import Mock, patch
 
 import pytest
 
-from kubeflow.common.testing import FAILED, SUCCESS, TestCase
 from kubeflow.trainer.backends.localprocess.backend import LocalProcessBackend
 from kubeflow.trainer.backends.localprocess.constants import LOCAL_RUNTIME_IMAGE
 from kubeflow.trainer.backends.localprocess.types import (
@@ -35,6 +34,7 @@ from kubeflow.trainer.options import (
     RuntimePatch,
 )
 from kubeflow.trainer.types import types
+from test.testing import FAILED, SUCCESS, TestCase
 
 # Test constants
 TORCH_RUNTIME = constants.DEFAULT_TRAINING_RUNTIME
