--- kubeflow/trainer/test/common.py
@@ -1,11 +1,0 @@
-# Shared test utilities and types for Kubeflow Trainer tests.
-# Common helpers have moved to kubeflow.common.testing.
-
-from kubeflow.common.testing import (  # noqa: F401
-    DEFAULT_NAMESPACE,
-    FAILED,
-    RUNTIME,
-    SUCCESS,
-    TIMEOUT,
-    TestCase,
-)
