+++ kubeflow/trainer/types/types_test.py
@@ -14,8 +14,8 @@
 
 import pytest
 
-from kubeflow.common.testing import FAILED, SUCCESS, TestCase
 from kubeflow.trainer.types import types
+from test.testing import FAILED, SUCCESS, TestCase
 
 
 @pytest.mark.parametrize(
